+++ src/lsapy/core/formatting.py
@@ -24,21 +24,24 @@ if TYPE_CHECKING:
 
 
 def sf_repr(sf: SuitabilityFunction) -> str:
-    """Return a string representation of a SuitabilityFunction."""
-    start = f"SuitabilityFunction(func={sf.func.__name__}"
-    if sf.params:
-        end = f", params={sf.params})"
-    else:
-        end = ")"
-    return f"{start}{end}"
-
-
-def sf_short_repr(sf: SuitabilityFunction) -> str:
     """Return a short string representation of a SuitabilityFunction."""
     func = f"{sf.func.__name__}"
     if not sf.params:
         return f"{func}()"
     return f"{func}({', '.join(f'{k}={v}' for k, v in sf.params.items())})"
+
+
+def sc_func_repr(func) -> str:
+    """Return a string representation of a suitability criteria standardization function."""
+    if isinstance(func, functools.partial):
+        f = f"{func.func.__name__}"
+        if not func.args and not func.keywords:
+            return f"{f}()"
+        args = [repr(a) for a in func.args] if func.args else []
+        kwargs = [f"{k}={v!r}" for k, v in func.keywords.items()] if func.keywords else []
+        return f"{f}({', '.join(args + kwargs)})"
+    else:
+        return repr(func)
 
 
 def sc_params_repr(sc: SuitabilityCriteria) -> str:
@@ -68,22 +71,32 @@ def sc_repr(sc: SuitabilityCriteria) -> str:
 
     col_width = _calculate_col_width([f"{k}:" for k in sc.attrs.keys()] + ["Dimensions"])
 
-    summary = [f"<SuitabilityCriteria> {sc.name!r}{sc_params_repr(sc)}"]
+    if not sc.name:
+        name = ""
+    else:
+        name = f" {sc.name!r}"
+
+    summary = [f"<SuitabilityCriteria>{name} {sc_params_repr(sc)}"]
 
     if sc.func:
-        summary.extend(["Function:", f"    {maybe_truncate(sf_repr(sc.func), max_width)}"])
+        summary.extend(["Function:", f"    {maybe_truncate(sc_func_repr(sc.func), max_width)}"])
 
-    dims = pretty_print("    Dimensions", col_width)
-    summary.extend(
-        [
-            "Indicator:",
-            f"{pretty_print('    Name', col_width)}{sc.indicator.name} ",
-            data_repr(sc.indicator, col_width, max_width),
-            f"{dims}{dim_summary_limited(sc.indicator.sizes, len(dims) + 1, max_rows)} ",
-        ]
-    )
+    if sc.indicator is not None:
+        dims = pretty_print("    Dimensions", col_width)
+        summary.extend(
+            [
+                "Indicator:",
+                f"{pretty_print('    Name', col_width)}{sc.indicator.name} ",
+                data_repr(sc.indicator, col_width, max_width),
+                f"{dims}{dim_summary_limited(sc.indicator.sizes, len(dims) + 1, max_rows)} ",
+            ]
+        )
 
-    summary.append(attrs_repr(sc.attrs, col_width=col_width, max_rows=max_rows))
+    if sc.attrs:
+        summary.append(attrs_repr(sc.attrs, col_width=col_width, max_rows=max_rows))
+
+    if not sc.func and not sc.indicator and not sc.attrs:
+        summary.append("    *undefined*")
 
     return "\n".join(summary)
 
@@ -113,7 +126,7 @@ def summarize_criteria(
         cat = ""
 
     if criteria.func:
-        func = f"{sf_short_repr(criteria.func)} "
+        func = f"{sc_func_repr(criteria.func)} "
     else:
         func = ""