+++ src/lsapy/criteria.py
@@ -2,13 +2,14 @@
 
 from __future__ import annotations
 
-from collections.abc import Mapping
+from collections.abc import Callable, Mapping
+from functools import partial
 from typing import Any
 
 import xarray as xr
 
 import lsapy.core.formatting as fmt
-from lsapy.functions import SuitabilityFunction
+from lsapy.core.functions import get_function_from_name
 
 __all__ = ["SuitabilityCriteria"]
 
@@ -27,8 +28,10 @@ class SuitabilityCriteria:
         Name of the suitability criteria.
     indicator : xr.DataArray
         Indicator on which the criteria is based.
-    func : SuitabilityFunction, optional
-        Suitability function describing how the suitability of the criteria is computed.
+    func : Callable, optional
+        Standardization function that takes the indicator as input and returns the suitability values.
+    fparams : dict, optional
+        A dictionary of parameters to pass to the function `func`. The default is None.
     weight : int | float, optional
         Weight of the criteria used in the aggregation process if a weighted aggregation method is used.
         The default is 1.
@@ -62,7 +65,8 @@ class SuitabilityCriteria:
     ...     weight=3,
     ...     category="soilTerrain",
     ...     indicator=drainage,
-    ...     func=SuitabilityFunction(name="discrete", params={"rules": {0: 0, 1: 0.1, 2: 0.5, 3: 0.9, 4: 1}}),
+    ...     func="discrete",
+    ...     fparams={"rules": {0: 0, 1: 0.1, 2: 0.5, 3: 0.9, 4: 1}},
     ... )
 
     Here is another example using the sample climate data with the growing degree days (GDD)
@@ -76,15 +80,17 @@ class SuitabilityCriteria:
     ...     weight=1,
     ...     category="climate",
     ...     indicator=gdd,
-    ...     func=SuitabilityFunction(name="vetharaniam2022_eq5", params={"a": -1.41, "b": 801}),
+    ...     func="vetharaniam2022_eq5",
+    ...     fparams={"a": -1.41, "b": 801},
     ... )
     """
 
     def __init__(
         self,
-        name: str,
-        indicator: xr.DataArray,
-        func: SuitabilityFunction | None = None,
+        name: str | None = None,
+        indicator: xr.DataArray | None = None,
+        func: Callable | None = None,
+        fparams: dict[str, Any] | None = None,
         weight: int | float | None = 1,
         category: str | None = None,
         long_name: str | None = None,
@@ -95,9 +101,12 @@ class SuitabilityCriteria:
     ) -> None:
         self.name = name
         self.indicator = indicator
-        self.func = func
         self.weight = weight
         self.category = category
+
+        if isinstance(func, str):
+            func = get_function_from_name(func)
+        self.func = partial(func, **fparams) if fparams else func
 
         self._attrs = {}
         if long_name:
@@ -110,11 +119,171 @@ class SuitabilityCriteria:
             self._attrs.update(attrs)
 
         self.is_computed = is_computed
-        self._from_indicator = _get_indicator_description(indicator)
 
     def __repr__(self) -> str:
         """Return a string representation of the suitability criteria."""
         return fmt.sc_repr(self)
+
+    @property
+    def name(self) -> str:
+        """
+        The name of the criteria.
+
+        Returns
+        -------
+        str
+            The name of the criteria.
+        """
+        return self._name
+
+    @name.setter
+    def name(self, value: str | None) -> None:
+        """
+        Set the name of the criteria.
+
+        Parameters
+        ----------
+        value : str | None
+            The name of the criteria to set.
+        """
+        self._name = value
+
+    @property
+    def indicator(self) -> xr.DataArray:
+        """
+        The indicator DataArray.
+
+        Returns
+        -------
+        xr.DataArray
+            The indicator DataArray.
+        """
+        return self._indicator
+
+    @indicator.setter
+    def indicator(self, value: xr.DataArray) -> None:
+        """
+        Set the indicator DataArray.
+
+        Parameters
+        ----------
+        value : xr.DataArray
+            The indicator DataArray to set.
+        """
+        if not isinstance(value, xr.DataArray) and value is not None:
+            raise TypeError("The indicator must be an xarray DataArray.")
+        if value is not None:
+            self._from_indicator = _get_indicator_description(value)
+        self._indicator = value
+
+    @property
+    def func(self) -> Callable | partial | None:
+        """
+        The standardization function.
+
+        Returns
+        -------
+        Callable | None
+            The standardization function.
+        """
+        return self._func
+
+    @func.setter
+    def func(self, value: Callable | partial | None) -> None:
+        """
+        Set the standardization function.
+
+        Parameters
+        ----------
+        value : Callable | None
+            The standardization function to set.
+        """
+        if not isinstance(value, Callable) and value is not None:
+            raise TypeError("The function must be a callable.")
+        self._func = value
+
+    @property
+    def weight(self) -> float:
+        """
+        The weight of the suitability criteria.
+
+        Returns
+        -------
+        float
+            The weight of the suitability criteria.
+        """
+        return self._weight
+
+    @weight.setter
+    def weight(self, value: int | float | None) -> None:
+        """
+        Set the weight of the suitability criteria.
+
+        Parameters
+        ----------
+        value : int | float | None
+            The weight of the suitability criteria. If None, the weight is set to 1.
+        """
+        if value is None:
+            self._weight = 1.0
+        elif not isinstance(value, (int, float)):
+            raise TypeError("The weight must be a number.")
+        elif value <= 0:
+            raise ValueError("The weight must be a positive number.")
+        else:
+            self._weight = float(value)
+
+    @property
+    def category(self) -> str | None:
+        """
+        The category of the suitability criteria.
+
+        Returns
+        -------
+        str | None
+            The category of the suitability criteria.
+        """
+        return self._category
+
+    @category.setter
+    def category(self, value: str | None) -> None:
+        """
+        Set the category of the suitability criteria.
+
+        Parameters
+        ----------
+        value : str | None
+            The category of the suitability criteria. If None, the category is set to None.
+        """
+        if value is not None and not isinstance(value, str):
+            raise TypeError("The category must be a string.")
+        self._category = value
+
+    @property
+    def is_computed(self) -> bool:
+        """
+        Whether the indicator data already contains the computed suitability values.
+
+        Returns
+        -------
+        bool
+            True if the indicator data already contains the computed suitability values, False otherwise.
+        """
+        return self._is_computed
+
+    @is_computed.setter
+    def is_computed(self, value: bool) -> None:
+        """
+        Set whether the indicator data already contains the computed suitability values.
+
+        Parameters
+        ----------
+        value : bool
+            True if the indicator data already contains the computed suitability values, False otherwise.
+        """
+        if not isinstance(value, bool):
+            raise TypeError("is_computed must be a boolean.")
+        self._is_computed = value
 
     @property
     def attrs(self) -> dict[Any, Any]: